simplify changes and remove duplicate code
